% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootpval.R
\name{bootpval}
\alias{bootpval}
\title{bootpval function}
\usage{
bootpval(x = x1, conf.level = 0.95, iter = 3000, mu0 = 0, test = "two")
}
\arguments{
\item{x}{vector of data values for which the p-value is calculated.}

\item{conf.level}{confidence level for the confidence interval (default is 0.95).}

\item{iter}{number of bootstrap iterations (default is 3000).}

\item{mu0}{null hypothesis value for the mean (default is 0).}

\item{test}{type of hypothesis test to perform.
Options are `"two"` (two-tailed), `"upper"` (upper tail), or `"lower"` (lower tail).}
}
\value{
A list containing the following:
  - `pvalue`: The calculated p-value based on the bootstrap resampling.
  - `tcalc`: The calculated t-statistic for the original data.
  - `n`: The sample size.
  - `x`: The original data vector.
  - `test`: The type of test performed.
  - `ci`: The confidence interval for the mean under the null hypothesis.
}
\description{
This function performs a bootstrap hypothesis test and calculates the p-value for a given dataset.
It can perform a two-sided, upper, or lower tail test and generates a histogram of the bootstrap resampled t-statistics.
}
\examples{
bootpval(x = x1, conf.level = 0.95, iter = 3000, mu0 = 5, test = "two")
}
