% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{myboot2 function}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{Number of bootstrap iterations (default: 10000).}

\item{x}{Sample data.}

\item{fun}{String or function specifying the statistic to compute (default: `"mean"`).}

\item{alpha}{Value between 0 and 1 specifying the significance level for the confidence interval (default: 0.05)}

\item{cx}{Size of text annotations on the plot (default: 1.5).}

\item{...}{Additional arguments passed to the `hist` function for customizing the histogram.}
}
\value{
A list (invisible) containing:
\itemize{
  \item \code{ci}: The bootstrap confidence interval as a numeric vector of length 2.
  \item \code{fun}: The statistic function used.
  \item \code{x}: The original sample data.
  \item \code{xstat}: The bootstrap sample statistics as a numeric vector.
}
}
\description{
This function performs a bootstrap resampling procedure to calculate confidence intervals
for a specified statistic (e.g., mean, median). It visualizes the distribution of the bootstrap sample
statistics with a histogram, including a confidence interval and point estimate annotations.
}
\examples{
myboot2(iter = 5000, x = c(2, 4, 7, 9, 11), fun = "mean", alpha = 0.05)
myboot2(iter = 10000, x = c(2, 4, 7, 9, 11), fun = "median", col = "blue")
}
