% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myci.R
\name{myci}
\alias{myci}
\title{myci function}
\usage{
myci(x, percent = 0.95, mode = "mean")
}
\arguments{
\item{x}{values to calculate the confidence interval.}

\item{percent}{confidence level (default is 0.95 for a 95% confidence interval).}

\item{mode}{the type of confidence interval to compute.
`"mean"` (default) for the confidence interval of the mean and `"variance"` for the confidence interval of the variance.}
}
\value{
A list containing the confidence interval (`myci`) and the confidence interval from a t-test (`ttest_ci`) for the mean.
For variance, only the CI is returned. If the `mode` is invalid, a message indicating the issue is returned.
}
\description{
This function calculates the confidence interval (CI) for either the mean or variance of a dataset `x`.
It can return the CI for the mean using the Student's t-distribution or for the variance using the Chi-square distribution.
The function also provides an option to conduct a t-test for comparison when calculating the CI for the mean.
}
\examples{
myci(x, percent = 0.95, mode = "mean")
}
