% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltp.R
\name{mycltp}
\alias{mycltp}
\title{mycltp function}
\usage{
mycltp(n, iter, lambda = 10, ...)
}
\arguments{
\item{n}{The sample size for each iteration.}

\item{iter}{Number of iterations (number of samples to draw).}

\item{lambda}{Rate parameter (\eqn{\lambda}) of the Poisson distribution (default: 10).}

\item{...}{Arguments to be passed to the `hist` function for customizing the histogram of sample means.}
}
\value{
None. This function produces multiple plots as side effects:
\itemize{
  \item A histogram of sample means with a theoretical normal curve overlay.
  \item A barplot of relative frequencies of sampled values.
  \item A plot of the Poisson probability mass function (PMF) for the given \eqn{\lambda}.
}
}
\description{
This function demonstrates the Central Limit Theorem (CLT) by simulating sample means
from a Poisson distribution. It produces a histogram of sample means with an overlay of
a theoretical normal curve, as well as visualizations of the original data distribution.
}
\examples{
mycltp(n = 30, iter = 100, lambda = 10)
mycltp(n = 50, iter = 200, lambda = 5, col = "lightblue", border = "blue")
}
