% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlikg.R
\name{mymaxlikg}
\alias{mymaxlikg}
\title{mymaxlikg function}
\usage{
mymaxlikg(lfun = "logbin2", theta)
}
\arguments{
\item{lfun}{A likelihood function (default: `"logbin2"`)}

\item{theta}{Values to evaluate the likelihood function over.}
}
\value{
The parameter (`theta`) that maximizes the likelihood.
}
\description{
This function performs a maximum likelihood estimation by evaluating a likelihood
across a range of parameter values (`theta`). It visualizes the likelihood function, highlights
the maximum likelihood estimate (MLE), and returns the parameter value corresponding to the MLE.
}
\examples{
logbin2 = function(theta) {
log(dbinom(2, prob = theta, size = 6)) + log(dbinom(4, prob = theta, size = 10))
}
mymaxlikg(loglik, seq(0.01, 0.99, by = 0.01))
}
