% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{myncurve function}
\usage{
myncurve(mu, sigma)
}
\arguments{
\item{mu}{Mean of the normal distribution.}

\item{sigma}{Standard deviation of the normal distribution.}
}
\value{
A list containing the parameters:
\itemize{
  \item \code{mu}: The mean of the normal distribution.
  \item \code{sigma}: The standard deviation of the normal distribution.
}
}
\description{
This function plots the probability density function (PDF) of a normal distribution
with specified mean (`mu`) and standard deviation (`sigma`). It also returns a list containing
the input parameters for reference.
}
\examples{
# Plot a normal curve with mean 0 and standard deviation 1
myncurve(mu = 0, sigma = 1)

}
