% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysample.R
\name{mysample}
\alias{mysample}
\title{mysample function}
\usage{
mysample(n, iter = 10, time = 0.5)
}
\arguments{
\item{n}{Size of the sample to draw in each iteration.}

\item{iter}{Number of iterations to perform.}

\item{time}{Pause duration (in seconds) between iterations.}
}
\value{
None.
}
\description{
This function generates random samples, converts them to factors,
and creates barplots visualising multiple iterations.
}
\examples{
mysample(n = 5)
mysample(n = 7, iter = 15, 0.5)

}
