% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{n tickets}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{Number of seats}

\item{gamma}{Probability of overbooking}

\item{p}{Probability of showing up}
}
\value{
Displays discrete and normal plots and prints the list of values
}
\description{
n tickets
}
