% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwcurve.R
\name{pwcurve}
\alias{pwcurve}
\title{pw.curve function}
\usage{
pwcurve(x, coef)
}
\arguments{
\item{x}{vector representing the independent variable.}

\item{coef}{vector of length 3 containing the coefficients for the piecewise curve.}
}
\value{
Calculate the piecewise curve
}
\description{
pw.curve function
}
\examples{
pw.curve(10, c(1, 2, 3)) # Example usage

}
